+++ lib/src/main/kotlin/net/dean/jraw/models/GeneralSort.kt
@@ -1,0 +1,6 @@
+package net.dean.jraw.models
+
+internal enum class GeneralSort(override val requiresTimePeriod: Boolean = false) : Sorting {
+    /** Used only internally as a default sort */
+    NEW
+}
