+++ lib/src/main/kotlin/net/dean/jraw/models/Sorting.kt
@@ -12,55 +12,3 @@ interface Sorting {
     val name: String
 }
 
-internal enum class GeneralSort(override val requiresTimePeriod: Boolean = false) : Sorting {
-    /** Used only internally as a default sort */
-    NEW
-}
-
-enum class SubredditSort(override val requiresTimePeriod: Boolean = false) : Sorting {
-    /** Popular posts, takes into account score, age, and other factors. */
-    HOT,
-
-    /** New posts */
-    NEW,
-
-    /** Somewhat new posts that have a lot of karma relative to others posted at the same time */
-    RISING,
-
-    /** Posts that received a lot of both upvotes and downvotes */
-    CONTROVERSIAL(true),
-
-    /** Posts that received the biggest score (upvotes minus downvotes) */
-    TOP(true);
-}
-
-enum class UserHistorySort(override val requiresTimePeriod: Boolean = false) : Sorting {
-    /** Popular posts and comments, takes into account score, age, and other factors. */
-    HOT,
-
-    /** New posts and comments */
-    NEW,
-
-    /** Posts and comments that received a lot of both upvotes and downvotes */
-    CONTROVERSIAL(true),
-
-    /** Posts and comments that received the biggest score (upvotes minus downvotes) */
-    TOP(true);
-}
-
-enum class SearchSort(override val requiresTimePeriod: Boolean = true) : Sorting {
-    /** Posts that match the search query the most */
-    RELEVANCE,
-
-    /** Popular posts, takes into account score, age, and other factors. */
-    HOT,
-
-    /** Posts that received the biggest score (upvotes minus downvotes) */
-    TOP,
-
-    /** New posts */
-    NEW,
-
-    /** Posts that received the most comments */
-    COMMENTS;
-}
